Create default parameter list

Create parameter object using default values.

create_default_par_list(
  models = NULL,
  mean_fun = NULL,
  mean_class = NULL,
  err_dist = NULL,
  err_class = NULL,
  binomial_n = NA,
  method = "paired"
)

Arguments

models

Models defined by set_models() object.

mean_fun

Mean function.

mean_class

String that indicates the class of mean functions to list.

err_dist

Error distribution.

err_class

String that indicates the class of error distributions to list.

binomial_n

Value of binomial n parameter for binomial models.

method

How mean function and error distributions should be combined; "paired" or "crossed".

Value

List of parameter simulation objects.

Examples

## Create default parameters from set_models() object
Models <- set_models (mean_fun=c("gaussian", "beta", "sech"),
                      err_dist=c("poisson","zip"), method="crossed")
Pars <- create_default_par_list (Models)

## Create parameter object by pairing mean function and error distribution
Pars <- create_default_par_list (method="paired",
                                 mean_fun=c("gaussian", "beta"), 
                                 err_dist=c("zip", "zinb"))

## Create parameter object by crossing mean function and error distribution
Pars <- create_default_par_list (method="crossed",
                                 mean_fun=c("gaussian", "beta"), 
                                 err_dist=c("poisson", "zip", "zinb"))