Create list of simulated data sets

Simulate data sets defined by the given parameter list.

create_simulated_datasets(
  Pars,
  x = NULL,
  N = 500,
  xmin = 0,
  xmax = 100,
  seed = NULL,
  echo = FALSE
)

Arguments

Pars
List containg parameters for the simulated data sets.

x
Vector of x values to simulated data. If not given, N, xmin, and xmax must be).

N
(optional) Sample size of data set to simulate. Default value 500.

xmin
(optional) Minimum value of x variable. Default value 0.

xmax
(optional) Maximum value of x variable. Default value 100.

seed
Random number seed.

echo
If TRUE display list describing data sets simulated

Value

Data frame containing one x variable and simulated data sets from Pars object.

Examples

## Set model
Models <- set_models (mean_fun="gaussian", err_dist="negbin")

## Set list of default parameters for each model
Pars <- create_default_par_list (Models)

## Simulate data
Data <- create_simulated_datasets (Pars, seed=12345)