Error distributions

Display error distributions of given class.

error_distributions(err_dist = NULL, err_class = NULL)

Arguments

err_dist
Vector of error distribution(s).

err_class
String that indicates the which error distributions to list.

"binary"

0/1 data; ("bernoulli").

"binomial"

Binomial data represented as counts or proportions; ("binomial.count", "binomial.prop").

"percentage"

[0-1] data; ("tab", "zitab").

"count"

Count data; ("poisson", "negbin", "zip", "zinb", "zipl", "zinbl", "zipl.mu", "zinbl.mu").

"abundance"

Non-negative continous data; ("gaussian", "tweedie", "zig", "zigl", "zigl.mu", "ziig", "ziigl", "ziigl.mu").

Value

If no arguments supplied, return a data frame listing all possible error distributions with corresponding error classes; if a vector of error classes is supplied, return corresponding error distributions; if a vector of error distributions is supplied, return corresponding error classes.

Examples

## Print data frame of all error distributions with corresponding error classes
error_distributions()

#>          err_dist  err_class
#> 1       bernoulli     binary
#> 2  binomial.count   binomial
#> 3   binomial.prop   binomial
#> 4         poisson      count
#> 5          negbin      count
#> 6             zip      count
#> 7            zinb      count
#> 8            zipl      count
#> 9           zinbl      count
#> 10        zipl.mu      count
#> 11       zinbl.mu      count
#> 12       gaussian  abundance
#> 13        tweedie  abundance
#> 14            zig  abundance
#> 15           zigl  abundance
#> 16        zigl.mu  abundance
#> 17           ziig  abundance
#> 18          ziigl  abundance
#> 19       ziigl.mu  abundance
#> 20            tab percentage
#> 21          zitab percentage

## Print all error distributions
error_distributions(err_class="all")

#>  [1] "bernoulli"      "binomial.count" "binomial.prop"  "poisson"       
#>  [5] "negbin"         "zip"            "zinb"           "zipl"          
#>  [9] "zinbl"          "zipl.mu"        "zinbl.mu"       "gaussian"      
#> [13] "tweedie"        "zig"            "zigl"           "zigl.mu"       
#> [17] "ziig"           "ziigl"          "ziigl.mu"       "tab"           
#> [21] "zitab"         

## Print error distributions for count data
error_distributions(err_class="count")

#> [1] "poisson"  "negbin"   "zip"      "zinb"     "zipl"     "zinbl"    "zipl.mu" 
#> [8] "zinbl.mu"

## Print error classes for poisson and gaussian data
error_distributions(err_dist=c("poisson", "gaussian"))

#> [1] "count"     "abundance"