Mean functions

Lists all possible mean functions of given class.

mean_functions(mean_fun = NULL, mean_class = NULL)

Arguments

mean_fun

Vector of mean function(s).

mean_class

String that indicates the class of mean functions to list. Possible values are:

"test"

Constant and uniform mean functions.

"main"

Main mean functions.

"sub"

Special cases of main mean functions.

"all"

All mean functions.

Value

Vector of mean functions of given class. If no arguments are supplied, return a data frame listing all possible mean functions with corresponding classes; if a vector of classes is supplied, return corresponding mean functions; if a vector of mean functions is supplied, return corresponding classes.

Examples

## Print data frame of all mean functions with corresponding classes
mean_functions()
#>             mean_fun mean_class
#> 1           constant       test
#> 2            uniform       test
#> 3               beta       main
#> 4               sech       main
#> 5           gaussian       main
#> 6        mixgaussian       main
#> 7               hofV       main
#> 8            sech.p1        sub
#> 9          sech.r0p1        sub
#> 10 mixgaussian.equal        sub
#> 11             hofII        sub
#> 12             hofIV        sub
#> 13            hofIVb        sub
#> 14             hofVb        sub

## Constant and uniform mean functions to test code
mean_functions(mean_class="test")
#> [1] "constant" "uniform" 

## Main mean functions
mean_functions(mean_class="main")
#> [1] "beta"        "sech"        "gaussian"    "mixgaussian" "hofV"       

## Classes of given mean functions
mean_functions(mean_fun=c("beta", "hofIV"))
#> [1] "main" "sub"