Fit species-environment non-linear model via maximum likelihood

'senlm' fits a species-environment non-linear model via maximum likelihood.

senlm(
  model = NULL,
  data = NULL,
  xvar = NULL,
  yvar = NULL,
  mean_fun = NULL,
  err_dist = NULL,
  binomial_n = NULL,
  y = NULL,
  x = NULL,
  conf.level = 0.95
)

Arguments

model

Model to fit.

data

Data frame of x and y.

xvar

Name of x (domain) variable (or column number).

yvar

Name of y (response) variable (or column number).

mean_fun

Mean function (if model not supplied).

err_dist

Error distribution (if model not supplied).

binomial_n

The binomial n parameter, if error distribution is "binomial.count" or "binomial.percent" (if model not supplied).

y

Repsonse variable vector (if data not supplied).

x

Vector of x (domain) values (if data not supplied).

conf.level

Confidence level for parameter confidence intervals. Default is 0.95.

Value

Object containg model fit to data y and x.

Examples

if (FALSE) {
## Simulate data

dat  <- create_simulated_datasets(pars, N=100, xmin=0, xmax=100, seed=12345)
## Fit model
fit <- senlm(mean_fun="gaussian", err_dist="poisson", data=dat,
             xvar="x", yvar="gaussian_poisson")

## Real data
model <- set_models(mean_fun="gaussian", err_dist="zip")
fit <- senlm(model=model, data=haul, xvar="depth", yvar="Sebastolobus.altivelis")
}